# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /me:
    get:
      operationId: getMe
      tags: [me]
      summary: Current authenticated user (identity + tier)
      responses:
        "200":
          description: Authenticated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MeResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /me/state:
    get:
      operationId: getMeState
      tags: [me]
      summary: Full personal state for the signed-in user (goals, prefs, draft)
      responses:
        "200":
          description: User state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MeStateResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: putMeState
      tags: [me]
      summary: Replace personal state with optimistic locking
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MeStateRequest"
      responses:
        "200":
          description: Updated state with new version
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MeStateResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Version conflict — fetch latest before retrying
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MeStateConflictResponse"
  /me/tier-history:
    get:
      operationId: getMeTierHistory
      tags: [me]
      summary: Subscription tier transition history for the signed-in user
      description: Returns every tier change in reverse-chronological order. Useful for support tooling to answer billing questions without database access.
      parameters:
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 50
          description: Maximum number of records to return (1–100, default 50).
      responses:
        "200":
          description: Tier transition history
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TierHistoryResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  securitySchemes:
    AdminKey:
      type: apiKey
      in: header
      name: X-Admin-Key
      description: Secret key required for all /admin/* endpoints. Set via the ADMIN_API_KEY environment variable on the server.
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
...
    TierTransitionEntry:
      type: object
      description: A single tier change event in the user's subscription history.
      properties:
        id:
          type: integer
        fromTier:
          type: string
        toTier:
          type: string
        triggeredBy:
          type: string
          description: '"webhook" or "sync-tier"'
        eventType:
          oneOf:
            - type: string
            - type: "null"
          description: Raw RevenueCat event type (e.g. "INITIAL_PURCHASE"). Null for sync-tier entries.
        createdAt:
          type: string
          format: date-time
      required: [id, fromTier, toTier, triggeredBy, eventType, createdAt]
    TierHistoryResponse:
      type: object
      properties:
        transitions:
          type: array
          items:
            $ref: "#/components/schemas/TierTransitionEntry"
        total:
          type: integer
          description: Total number of transitions returned in this response.
      required: [transitions, total]
    MeResponse:
      type: object
      description: Public-safe identity and tier for the signed-in user.
      properties:
        authUserId:
          type: string
        email:
          oneOf:
            - type: string
            - type: "null"
        tier:
          type: string
          description: Server-side subscription tier. "free" today; later "pro" etc.
      required: [authUserId, email, tier]
    MeStateGoalBlob:
      type: object
      description: Opaque per-goal blob (profile, roadmap, history, plans, reflections, learnedProfile, evolutions, coachMemory, etc). Schema enforced client-side.
      additionalProperties: true
    MeStateAccountPrefsBlob:
      type: object
      description: Opaque account preferences blob. Schema enforced client-side.
      additionalProperties: true
    MeStateDraftBlob:
      type: object
      description: Opaque pending intake draft blob. Schema enforced client-side.
      additionalProperties: true
    MeStateResponse:
      type: object
      properties:
        authUserId:
          type: string
        email:
          oneOf:
            - type: string
            - type: "null"
        tier:
          type: string
        goals:
          type: array
          items:
            $ref: "#/components/schemas/MeStateGoalBlob"
        activeGoalId:
          oneOf:
            - type: string
            - type: "null"
        accountPrefs:
          $ref: "#/components/schemas/MeStateAccountPrefsBlob"
        pendingDraft:
          oneOf:
            - $ref: "#/components/schemas/MeStateDraftBlob"
            - type: "null"
        version:
          type: integer
          description: Monotonic version used for optimistic concurrency. Bumped on every successful PUT.
      required: [authUserId, email, tier, goals, activeGoalId, accountPrefs, pendingDraft, version]
    MeStateRequest:
      type: object
      properties:
        goals:
          type: array
          items:
            $ref: "#/components/schemas/MeStateGoalBlob"
        activeGoalId:
          oneOf:
            - type: string
            - type: "null"
        accountPrefs:
          $ref: "#/components/schemas/MeStateAccountPrefsBlob"
        pendingDraft:
          oneOf:
            - $ref: "#/components/schemas/MeStateDraftBlob"
            - type: "null"
        expectedVersion:
          type: integer
          description: The version the client believes is current. If it doesn't match, the server returns 409 with the latest state.
      required: [goals, activeGoalId, accountPrefs, pendingDraft, expectedVersion]
    MeStateConflictResponse:
      type: object
      properties:
        error:
          type: string
        latest:
          $ref: "#/components/schemas/MeStateResponse"
      required: [error, latest]
...